+++ example/lib/main.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:oktoast/oktoast.dart';
 
 import 'common/tu_chong_repository.dart';
+import 'custom_toolbar.dart';
 
 void main() {
   runApp(MyApp());
@@ -43,9 +44,10 @@ class _MyHomePageState extends State<MyHomePage> {
   TuChongRepository _listSourceRepository = TuChongRepository();
   @override
   void initState() {
-    // TODO: implement initState
     pages.add(Page(
         PageType.text, "build special text and inline image in text field"));
+    pages.add(Page(
+        PageType.customToolBar, "custom selection tool bar for text field"));
 
     super.initState();
   }
@@ -53,14 +55,11 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   void dispose() {
     _listSourceRepository.dispose();
-    // TODO: implement dispose
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    // TODO: implement build
-
     var content = ListView.builder(
       itemBuilder: (_, int index) {
         var page = pages[index];
@@ -88,7 +87,10 @@ class _MyHomePageState extends State<MyHomePage> {
             onTap: () {
               switch (page.type) {
                 case PageType.text:
-                  pageWidget = new TextDemo(_listSourceRepository);
+                  pageWidget = TextDemo(_listSourceRepository);
+                  break;
+                case PageType.customToolBar:
+                  pageWidget = CustomToolBar();
                   break;
                 default:
                   break;
@@ -116,4 +118,4 @@ class Page {
   Page(this.type, this.description);
 }
 
-enum PageType { text }
+enum PageType { text, customToolBar }