+++ example/lib/special_text/my_special_text_span_builder.dart
@@ -6,11 +6,43 @@ import 'package:extended_text_library/extended_text_library.dart';
 import 'package:flutter/material.dart';
 
 class MySpecialTextSpanBuilder extends SpecialTextSpanBuilder {
+  Map<TextRange, TextStyle> specialTextStyle = Map<TextRange, TextStyle>();
+
   /// whether show background for @somebody
   final bool showAtBackground;
   final BuilderType type;
   MySpecialTextSpanBuilder(
       {this.showAtBackground: false, this.type: BuilderType.extendedText});
+
+  @override
+  TextSpan build(String data, {TextStyle textStyle, onTap}) {
+    TextSpan result = super.build(data, textStyle: textStyle, onTap: onTap);
+    handleSpeicalTextStyle(result);
+    return result;
+  }
+
+  TextSpan handleSpeicalTextStyle(TextSpan result) {
+    if (specialTextStyle.length != 0 &&
+        result != null &&
+        result.children != null) {
+      int index = 0;
+      List<InlineSpan> inlineList = new List<InlineSpan>();
+      for (InlineSpan item in result.children) {
+        if (item is SpecialInlineSpanBase) {
+          var base = item as SpecialInlineSpanBase;
+
+          index = base.end;
+        } else {
+          var start = index;
+          var end = index + item.toPlainText().length;
+
+          index = end;
+        }
+      }
+      return TextSpan(style: result.style, children: inlineList);
+    }
+    return result;
+  }
 
   @override
   SpecialText createSpecialText(String flag,
@@ -38,3 +70,7 @@ class MySpecialTextSpanBuilder extends SpecialTextSpanBuilder {
 }
 
 enum BuilderType { extendedText, extendedTextField }
+
+class SpecialTextStyle {
+  TextRange textRange;
+}