+++ lib/src/extended_text_field.dart
@@ -172,7 +172,8 @@ class ExtendedTextField extends StatefulWidget {
       this.buildCounter,
       this.scrollController,
       this.scrollPhysics,
-      this.specialTextSpanBuilder})
+      this.specialTextSpanBuilder,
+      this.textSelectionControls})
       : assert(textAlign != null),
         assert(readOnly != null),
         assert(autofocus != null),
@@ -198,6 +199,10 @@ class ExtendedTextField extends StatefulWidget {
         keyboardType = keyboardType ??
             (maxLines == 1 ? TextInputType.text : TextInputType.multiline),
         super(key: key);
+
+  /// An interface for building the selection UI, to be provided by the
+  /// implementor of the toolbar widget or handle widget
+  final TextSelectionControls textSelectionControls;
 
   ///build your ccustom text span
   final SpecialTextSpanBuilder specialTextSpanBuilder;
@@ -968,7 +973,7 @@ class _ExtendedTextFieldState extends State<ExtendedTextField>
       formatters.add(LengthLimitingTextInputFormatter(widget.maxLength));
 
     bool forcePressEnabled;
-    TextSelectionControls textSelectionControls;
+    TextSelectionControls textSelectionControls = widget.textSelectionControls;
     bool paintCursorAboveText;
     bool cursorOpacityAnimates;
     Offset cursorOffset;
@@ -978,7 +983,7 @@ class _ExtendedTextFieldState extends State<ExtendedTextField>
     switch (themeData.platform) {
       case TargetPlatform.iOS:
         forcePressEnabled = true;
-        textSelectionControls = cupertinoTextSelectionControls;
+        textSelectionControls ??= cupertinoExtendedTextSelectionControls;
         paintCursorAboveText = true;
         cursorOpacityAnimates = true;
         cursorColor ??= CupertinoTheme.of(context).primaryColor;
@@ -997,7 +1002,7 @@ class _ExtendedTextFieldState extends State<ExtendedTextField>
       case TargetPlatform.android:
       case TargetPlatform.fuchsia:
         forcePressEnabled = false;
-        textSelectionControls = materialTextSelectionControls;
+        textSelectionControls ??= materialExtendedTextSelectionControls;
         paintCursorAboveText = false;
         cursorOpacityAnimates = false;
         cursorColor ??= themeData.cursorColor;